Collection naming clean-up
